restore: Make snapshot list human-readable
When restoring from snapshots we should print something that is easy to
understand for humans, not ::registry::snapshot12.

Printing the snapshot ID is required for the --snapshot-id parameter of
restore, but putting it at the front of the line would potentially
confuse users since the selection code also prints numbers, so put it at
the end of the line.